Fuse Scene.hpp & Entity.hpp because of loop constrains
