+++ include/Voxymore/Components/CustomComponent.hpp
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "Voxymore/Core/YamlHelper.hpp"
-#include "Voxymore/Scene/Entity.hpp"
+#include "Voxymore/Scene/Scene.hpp"
 #include "imgui.h"
 #include <string>
 #include <vector>
