+++ src/Physics/Components/AnchoredSpringComponent.cpp
@@ -5,7 +5,7 @@
 #include "Voxymore/Physics/Components/AnchoredSpringComponent.hpp"
 #include "Voxymore/ImGUI/ImGuiLib.hpp"
 #include "Voxymore/Utils/Platform.hpp"
-#include "Voxymore/Scene/Entity.hpp"
+#include "Voxymore/Scene/Scene.hpp"
 
 #define DeserializeField(node, fieldVariable, fieldName, type, defaultValue)														\
 auto VXM_COMBINE(fieldVariable, Node) = node[fieldName]; 																			\
